requirements.txt
Django==2.2.13
django_filter==22.1
djangorestframework==3.14.0
drf_yasg==1.17.0
kafka==1.3.5
PyMySQL==0.9.3
ruamel.base==1.0.0
